+++ crates/teloxide-core/src/adaptors/parse_mode.rs
@@ -290,7 +290,11 @@ where
         set_game_score_inline,
         get_game_high_scores,
         approve_chat_join_request,
-        decline_chat_join_request
+        decline_chat_join_request,
+        get_available_gifts,
+        send_gift,
+        set_user_emoji_status,
+        save_prepared_inline_message
         => fid, ftyid
     }
 }